@@ -1,88 +1,111 @@
 {
-  "name": "@scaffold-eth/react-app",
-  "version": "1.1.0",
-  "homepage": ".",
-  "browserslist": {
-    "production": [
-      ">0.2% or last 2 versions",
-      "not dead"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "dependencies": {
-    "@davatar/react": "^1.8.1",
-    "@emotion/react": "^11.5.0",
-    "@emotion/styled": "^11.3.0",
-    "@ethersproject/providers": "^5.0.12",
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.58",
-    "@mui/material": "^5.1.0",
-    "@netlify/functions": "^0.8.0",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@walletconnect/web3-provider": "^1.6.5",
-    "axios": "^0.21.2",
-    "classnames": "^2.3.1",
-    "ethers": "^5.4.0",
-    "from-exponential": "^1.1.1",
-    "i18next": "^21.5.4",
-    "i18next-browser-languagedetector": "^6.1.2",
-    "intl": "^1.2.5",
-    "lodash": "^4.17.21",
-    "node-sass": "^6.0.1",
-    "node-watch": "^0.7.1",
-    "notistack": "^1.0.10",
-    "numeral": "^2.0.6",
-    "react": "^16.14.0",
-    "react-copy-to-clipboard": "^5.0.4",
-    "react-dom": "^16.14.0",
-    "react-i18next": "^11.14.3",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "react-snowflakes": "^1.0.0",
-    "react-snowstorm": "^0.0.3",
-    "redux": "^4.1.0",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
-    "sass": "^1.35.1",
-    "typescript": "^4.3.2",
-    "web3modal": "^1.9.4"
-  },
-  "devDependencies": {
-    "@types/lodash": "^4.14.172",
-    "@types/node": "^14.17.32",
-    "@types/numeral": "^2.0.2",
-    "@types/react": "^17.0.15",
-    "@types/react-copy-to-clipboard": "^5.0.2",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.1.7",
-    "@typescript-eslint/eslint-plugin": "^4.28.5",
-    "@typescript-eslint/parser": "^4.28.5",
-    "eslint": "^7.31.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "prettier": "^2.0.5"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "scripts": {
-    "build": "GENERATE_SOURCEMAP=false react-scripts build",
-    "eject": "react-scripts eject",
-    "start": "react-scripts start",
-    "test": "react-scripts test",
-    "lint": "npx eslint src/ --ext .ts,.js,.tsx,.jsx",
-    "lint-fix": "npx eslint src/ --ext .ts,.js,.tsx,.jsx --fix"
-  }
+    "name": "@scaffold-eth/react-app",
+    "version": "1.1.0",
+    "homepage": ".",
+    "browserslist": {
+        "production": [
+            ">0.2% or last 2 versions",
+            "not dead"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    },
+    "dependencies": {
+        "@emotion/react": "^11.9.0",
+        "@emotion/styled": "^11.8.1",
+        "@mui/icons-material": "^5.5.1",
+        "@mui/material": "^5.5.2",
+        "@reduxjs/toolkit": "^1.6.1",
+        "@walletconnect/web3-provider": "^1.6.5",
+        "axios": "^0.21.2",
+        "decimal.js": "^10.3.1",
+        "ethers": "^5.4.0",
+        "from-exponential": "^1.1.1",
+        "i18next": "^21.5.4",
+        "i18next-browser-languagedetector": "^6.1.2",
+        "lodash": "^4.17.21",
+        "luxon": "^2.3.1",
+        "node-sass": "^6.0.1",
+        "notistack": "^2.0.4",
+        "react": "^16.14.0",
+        "react-copy-to-clipboard": "^5.0.4",
+        "react-dom": "^16.14.0",
+        "react-i18next": "^11.14.3",
+        "react-redux": "^7.2.4",
+        "react-router-dom": "^5.2.0",
+        "react-scripts": "^4.0.3",
+        "redux": "^4.1.0",
+        "redux-logger": "^3.0.6",
+        "redux-thunk": "^2.3.0",
+        "reselect": "^4.1.5",
+        "sass": "^1.35.1",
+        "web3modal": "https://github.com/drtex0/web3modal-optimized"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.17.5",
+        "@babel/preset-env": "^7.16.11",
+        "@babel/preset-react": "^7.16.7",
+        "@testing-library/jest-dom": "^5.16.3",
+        "@testing-library/react": "^12.1.4",
+        "@testing-library/react-hooks": "^8.0.0",
+        "@types/chai": "^4.3.0",
+        "@types/lodash": "^4.14.172",
+        "@types/luxon": "^2.3.1",
+        "@types/node": "^14.18.12",
+        "@types/react": "^18.0.6",
+        "@types/react-copy-to-clipboard": "^5.0.2",
+        "@types/react-dom": "^17.0.9",
+        "@types/react-router-dom": "^5.1.7",
+        "@types/redux-mock-store": "^1.0.3",
+        "@typescript-eslint/eslint-plugin": "^5.13.0",
+        "@typescript-eslint/parser": "^5.0.0",
+        "babel-plugin-transform-imports": "^2.0.0",
+        "cross-env": "^7.0.3",
+        "eslint": "^7.31.0",
+        "eslint-config-airbnb": "^18.2.0",
+        "eslint-config-airbnb-typescript": "^17.0.0",
+        "eslint-config-prettier": "^6.11.0",
+        "eslint-plugin-babel": "^5.3.1",
+        "eslint-plugin-import": "^2.22.1",
+        "eslint-plugin-jest": "^26.1.5",
+        "eslint-plugin-jsx-a11y": "^6.4.1",
+        "eslint-plugin-prettier": "^3.1.4",
+        "eslint-plugin-react": "^7.22.0",
+        "eslint-plugin-react-hooks": "^4.2.0",
+        "node-watch": "^0.7.3",
+        "prettier": "^2.0.5",
+        "redux-mock-store": "^1.5.4",
+        "ts-node": "^10.7.0",
+        "typescript": "^4.6.2"
+    },
+    "resolutions": {
+        "web3modal/react-dom": "^16.14.0"
+    },
+    "eslintConfig": {
+        "extends": "react-app"
+    },
+    "scripts": {
+        "build": "react-scripts build",
+        "eject": "react-scripts eject",
+        "start": "react-scripts start",
+        "test": "react-scripts test ",
+        "lint": "npx eslint src/ --ext .ts,.js,.tsx,.jsx",
+        "lint-fix": "npx eslint src/ --ext .ts,.js,.tsx,.jsx --fix",
+        "prettier": "npx prettier --write 'src/**/*.{ts,tsx,js,jsx}'"
+    },
+    "jest": {
+        "collectCoverageFrom": [
+            "<rootDir>/src/**/*.{js,ts,jsx,tsx}",
+            "!<rootDir>/src/**/*.d.{js,ts,jsx,tsx}",
+            "!<rootDir>/src/locales/*.{js,ts,jsx,tsx}",
+            "!<rootDir>/src/index.{js,ts,jsx,tsx}",
+            "!<rootDir>/scripts/*",
+            "!<rootDir>/src/config/*",
+            "!<rootDir>/src/store/store.ts",
+            "!<rootDir>/src/tests/utils.tsx"
+        ]
+    }
 }
