@@ -1,20 +1,20 @@
-import { Box } from "@material-ui/core";
-import { Bond } from "../helpers/bond/bond";
+import { Box } from '@mui/material';
 
 interface IBondLogoProps {
-    bond: Bond;
+    bondLogoPath: string;
+    isLP: boolean;
 }
 
-function BondLogo({ bond }: IBondLogoProps) {
-    let style = { height: "32px", width: "32px" };
+function BondLogo({ bondLogoPath, isLP }: IBondLogoProps) {
+    let style = { height: '32px', width: '32px' };
 
-    if (bond.isLP) {
-        style = { height: "32px", width: "62px" };
+    if (isLP) {
+        style = { height: '32px', width: '62px' };
     }
 
     return (
-        <Box display="flex" alignItems="center" justifyContent="center" width={"64px"}>
-            <img src={bond.bondIconSvg} style={style} />
+        <Box display="flex" alignItems="center" justifyContent="center" width={'64px'}>
+            <img src={bondLogoPath} style={style} />
         </Box>
     );
 }
