@@ -1,23 +1,31 @@
-import React, { useEffect, useState } from "react";
-import App from "./App";
-import { HashRouter } from "react-router-dom";
-import { loadTokenPrices } from "../helpers";
-import Loading from "../components/Loader";
+import { lazy, Suspense } from 'react';
 
-function Root() {
-    const [loading, setLoading] = useState(true);
+import { BrowserRouter } from 'react-router-dom';
 
-    useEffect(() => {
-        loadTokenPrices().then(() => setLoading(false));
-    }, []);
+import Loader from 'components/Loader';
+import { useWeb3ContextInitialized } from 'contexts/web3/web3.hooks';
+import ViewBase from 'layout/ViewBase';
 
-    if (loading) return <Loading />;
+const App = lazy(() => import('./App'));
+
+function Root(): JSX.Element {
+    const web3ContextReady = useWeb3ContextInitialized();
+
+    if (!web3ContextReady) return <Loader />;
 
     return (
-        <HashRouter>
+        <Suspense fallback={<Loader />}>
             <App />
-        </HashRouter>
+        </Suspense>
     );
 }
 
-export default Root;
+// DOC : Using BrowserRouter before viewbase because of the <Link> components in Menu
+export const RenderWithViewBase = () => (
+    <BrowserRouter>
+        <ViewBase>
+            <Root />
+        </ViewBase>
+    </BrowserRouter>
+);
+export default RenderWithViewBase;
