@@ -1,60 +1,55 @@
-import { useEffect, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { useWeb3Context } from "../../../hooks";
-import { DEFAULD_NETWORK } from "../../../constants";
-import { IReduxState } from "../../../store/slices/state.interface";
-import { IPendingTxn } from "../../../store/slices/pending-txns-slice";
-import "./connect-menu.scss";
-import CircularProgress from "@material-ui/core/CircularProgress";
-import { colors } from "@material-ui/core";
+import { useCallback, useContext, useEffect, useState } from 'react';
 
-import { useTranslation } from "react-i18next";
+import { Box, Button } from '@mui/material';
+import { useTranslation } from 'react-i18next';
+import { useSelector } from 'react-redux';
+
+import { theme } from 'constants/theme';
+import { Web3Context } from 'contexts/web3/web3.context';
+import { useSignerConnected, useGoodNetworkCheck } from 'contexts/web3/web3.hooks';
+import { selectPendingTransactions } from 'store/modules/transactions/transactions.selectors';
 
 function ConnectMenu() {
     const { t } = useTranslation();
-    const { connect, disconnect, connected, web3, providerChainID, checkWrongNetwork } = useWeb3Context();
-    const dispatch = useDispatch();
-    const [isConnected, setConnected] = useState(connected);
+    const { memoConnect, memoDisconnect, state } = useContext(Web3Context);
 
-    let pendingTransactions = useSelector<IReduxState, IPendingTxn[]>(state => {
-        return state.pendingTransactions;
-    });
+    const isUserSigned = useSignerConnected();
 
-    let buttonText = t("ConnectWallet");
-    let clickFunc: any = connect;
-    let buttonStyle = {};
+    const isUserOnGoodNetwork = useGoodNetworkCheck();
 
-    if (isConnected) {
-        buttonText = t("Disconnect");
-        clickFunc = disconnect;
-    }
+    const pendingTransactions = useSelector(selectPendingTransactions);
 
-    if (pendingTransactions && pendingTransactions.length > 0) {
-        buttonText = t("CountPending", { count: pendingTransactions.length });
-        clickFunc = () => {};
-    }
+    const handleButtonClick = useCallback(() => {
+        return isUserSigned ? memoDisconnect(state.signer) : memoConnect();
+    }, [isUserSigned, memoConnect, memoDisconnect]);
 
-    if (isConnected && providerChainID !== DEFAULD_NETWORK) {
-        buttonText = t("WrongNetwork");
-        buttonStyle = { backgroundColor: "rgb(255, 67, 67)", color: "#ffffff" };
-        clickFunc = () => {
-            checkWrongNetwork();
-        };
-    }
+    const [buttonText, setButtonText] = useState(t('ConnectWallet'));
 
     useEffect(() => {
-        setConnected(connected);
-    }, [web3, connected]);
+        if (isUserSigned) {
+            if (!isUserOnGoodNetwork) {
+                setButtonText(t('WrongNetwork'));
+            } else if (pendingTransactions.length > 0) {
+                setButtonText(t('CountPending', { count: pendingTransactions.length })); // Ususally user can't have more than 1
+            } else {
+                setButtonText(t('Disconnect'));
+            }
+        }
+    }, [isUserSigned, pendingTransactions, isUserOnGoodNetwork]);
 
     return (
-        <div className="connect-button" style={buttonStyle} onClick={clickFunc}>
-            <p>{buttonText}</p>
-            {pendingTransactions.length > 0 && (
-                <div className="connect-button-progress">
-                    <CircularProgress size={15} color="inherit" />
-                </div>
-            )}
-        </div>
+        <Box>
+            <Button
+                sx={{
+                    background: isUserSigned && !isUserOnGoodNetwork ? 'red !important' : 'rgba(255, 255, 255, 0.9)',
+                    boxShadow: '0px 0px 10px rgba(44, 39, 109, 0.1)',
+                    padding: theme.spacing(1),
+                }}
+                onClick={handleButtonClick}
+            >
+                <>{buttonText}</>
+            </Button>
+        </Box>
     );
 }
 
