@@ -1,7 +1,7 @@
-import { Networks } from "../../constants/blockchain";
+import { Networks } from '../../constants/blockchain';
 
 export enum BondType {
-    StableAsset,
+    STABLE_ASSET,
     LP,
 }
 
@@ -13,4 +13,5 @@ export interface BondAddresses {
 export interface NetworkAddresses {
     [Networks.MAINNET]: BondAddresses;
     [Networks.RINKEBY]: BondAddresses;
+    [Networks.LOCAL]: BondAddresses;
 }
