@@ -1,10 +1,13 @@
-import fromExponential from "from-exponential";
+import fromExponential from 'from-exponential';
 
-export const trim = (number: number = 0, precision?: number) => {
-    const array = fromExponential(number).split(".");
-    if (array.length === 1) return fromExponential(number);
-    //@ts-ignore
-    array.push(array.pop().substring(0, precision));
-    const trimmedNumber = array.join(".");
-    return trimmedNumber;
+export const trim = (number = 0, precision?: number) => {
+    const [exp, decimals] = fromExponential(number).split('.');
+
+    if (!decimals) return exp;
+
+    return [exp, decimals.substring(0, precision)].join('.');
+};
+
+export const trimString = (numberInGWEI: string) => {
+    return Number(numberInGWEI).toFixed(2);
 };
