@@ -1,19 +1,25 @@
-import { ethers } from "ethers";
-import { useEffect, useState } from "react";
+import { useEffect, useState } from 'react';
 
-const useENS = (address: string) => {
+import { utils } from 'ethers';
+
+import { useWeb3Context } from 'contexts/web3/web3.context';
+
+const useENS = () => {
     const [ensName, setENSName] = useState<string | null>(null);
+
+    const {
+        state: { provider, signerAddress },
+    } = useWeb3Context();
 
     useEffect(() => {
         const resolveENS = async () => {
-            if (ethers.utils.isAddress(address)) {
-                const provider = new ethers.providers.JsonRpcProvider("https://rinkeby.infura.io/v3/254fc898c6c24be99475e8ec90ced016");
-                const ensName = await provider.lookupAddress(address);
-                setENSName(ensName);
+            if (signerAddress && provider && utils.isAddress(signerAddress)) {
+                const name = await provider.lookupAddress(signerAddress);
+                setENSName(name);
             }
         };
         resolveENS();
-    }, [address]);
+    }, [provider, signerAddress]);
 
     return { ensName };
 };
