@@ -1,27 +1,40 @@
-import ReactDOM from "react-dom";
-import Root from "./Root";
-import store from "./store/store";
-import { Provider } from "react-redux";
-import { Web3ContextProvider } from "./hooks";
-import { SnackbarProvider } from "notistack";
-import SnackMessage from "./components/Messages/snackbar";
-import "./i18n";
+import './i18n';
+
+import { CssBaseline } from '@mui/material';
+import { ThemeProvider } from '@mui/material/styles';
+import { SnackbarProvider } from 'notistack';
+import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+
+import { BSnackBar } from 'components/Messages/snackbar';
+import { theme } from 'constants/theme';
+import { Web3ContextProvider } from 'contexts/web3/web3.context';
+
+import Root from './Root';
+import store from './store/store';
 
 ReactDOM.render(
-    <SnackbarProvider
-        maxSnack={4}
-        anchorOrigin={{
-            vertical: "bottom",
-            horizontal: "right",
-        }}
-        content={(key, message: string) => <SnackMessage id={key} message={JSON.parse(message)} />}
-        autoHideDuration={10000}
-    >
-        <Provider store={store}>
+    <>
+        <CssBaseline />
+        <ThemeProvider theme={theme}>
             <Web3ContextProvider>
-                <Root />
+                <Provider store={store}>
+                    <SnackbarProvider
+                        maxSnack={3}
+                        anchorOrigin={{
+                            vertical: 'bottom',
+                            horizontal: 'left',
+                        }}
+                        autoHideDuration={5000}
+                        content={() => {
+                            return <BSnackBar severity="info" description="test" />;
+                        }}
+                    >
+                        <Root />
+                    </SnackbarProvider>
+                </Provider>
             </Web3ContextProvider>
-        </Provider>
-    </SnackbarProvider>,
-    document.getElementById("root"),
+        </ThemeProvider>
+    </>,
+    document.getElementById('root'),
 );
