@@ -1,42 +1,42 @@
 /// <reference types="node" />
-/// <reference types="react" />
+/// <reference types="react-scripts" />
 /// <reference types="react-dom" />
 
 declare namespace NodeJS {
-  interface ProcessEnv {
-    readonly NODE_ENV: 'development' | 'production' | 'test';
-    readonly PUBLIC_URL: string;
-  }
+    interface ProcessEnv {
+        readonly NODE_ENV: 'development' | 'production' | 'test';
+        readonly PUBLIC_URL: string;
+    }
 }
 
 declare module '*.avif' {
-  const src: string;
-  export default src;
+    const src: string;
+    export default src;
 }
 
 declare module '*.bmp' {
-  const src: string;
-  export default src;
+    const src: string;
+    export default src;
 }
 
 declare module '*.gif' {
-  const src: string;
-  export default src;
+    const src: string;
+    export default src;
 }
 
 declare module '*.jpg' {
-  const src: string;
-  export default src;
+    const src: string;
+    export default src;
 }
 
 declare module '*.jpeg' {
-  const src: string;
-  export default src;
+    const src: string;
+    export default src;
 }
 
 declare module '*.png' {
-  const src: string;
-  export default src;
+    const src: string;
+    export default src;
 }
 
 declare module '*.webp' {
@@ -45,27 +45,25 @@ declare module '*.webp' {
 }
 
 declare module '*.svg' {
-  import * as React from 'react';
+    import * as React from 'react';
 
-  export const ReactComponent: React.FunctionComponent<React.SVGProps<
-    SVGSVGElement
-  > & { title?: string }>;
+    export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement> & { title?: string }>;
 
-  const src: string;
-  export default src;
+    const src: string;
+    export default src;
 }
 
 declare module '*.module.css' {
-  const classes: { readonly [key: string]: string };
-  export default classes;
+    const classes: { readonly [key: string]: string };
+    export default classes;
 }
 
 declare module '*.module.scss' {
-  const classes: { readonly [key: string]: string };
-  export default classes;
+    const classes: { readonly [key: string]: string };
+    export default classes;
 }
 
 declare module '*.module.sass' {
-  const classes: { readonly [key: string]: string };
-  export default classes;
+    const classes: { readonly [key: string]: string };
+    export default classes;
 }
