@@ -1,15 +1,33 @@
-import "./notfound.scss";
+import { Box, Typography } from '@mui/material';
+import { useTranslation } from 'react-i18next';
+import { useSelector } from 'react-redux';
 
-import { useTranslation } from "react-i18next";
+import Loader from 'components/Loader';
+import { theme } from 'constants/theme';
+import { selectBondInstances } from 'store/modules/bonds/bonds.selector';
 
-function NotFound() {
+const NotFound = () => {
     const { t } = useTranslation();
 
+    const bondsReady = useSelector(selectBondInstances);
+
     return (
-        <div className="not-found-page">
-            <p>{t("PageNotFound")}</p>
-        </div>
+        <Box
+            sx={{
+                justifyContent: 'center',
+                alignItems: 'center',
+                display: 'flex',
+                height: '100%',
+            }}
+        >
+            {bondsReady.length === 0 && <Loader />}
+            {bondsReady.length > 0 && (
+                <Typography variant="h1" sx={{ textAlign: 'center', fontWeight: '600', color: theme.palette.primary.main }}>
+                    <>{t('PageNotFound')}</>
+                </Typography>
+            )}
+        </Box>
     );
-}
+};
 
 export default NotFound;
