@@ -1,16 +1,20 @@
-import { Box, Modal, Paper, SvgIcon, IconButton, FormControl, OutlinedInput, InputLabel, InputAdornment } from "@material-ui/core";
-import { useEffect, useState } from "react";
-import { ReactComponent as XIcon } from "../../assets/icons/x.svg";
-import "./bondSettings.scss";
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { useEffect, useState } from 'react';
 
-import { useTranslation } from "react-i18next";
+import { Box, Modal, Paper, SvgIcon, IconButton, FormControl, OutlinedInput, InputLabel, InputAdornment } from '@mui/material';
+import { useTranslation } from 'react-i18next';
+
+import { ReactComponent as XIcon } from '../../assets/icons/x.svg';
+// import './bondSettings.scss';
 
 interface IAdvancedSettingsProps {
     open: boolean;
     handleClose: () => void;
     slippage: number;
     recipientAddress: string;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     onRecipientAddressChange: (e: any) => void;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     onSlippageChange: (e: any) => void;
 }
 
@@ -36,11 +40,15 @@ function AdvancedSettings({ open, handleClose, slippage, recipientAddress, onRec
                     </IconButton>
                 </div>
 
-                <p className="hades-title">{t("bond:Settings")}</p>
+                <p className="hades-title">
+                    <>{t('bond:Settings')}</>
+                </p>
 
                 <Box className="card-content">
                     <InputLabel htmlFor="slippage">
-                        <p className="input-lable">{t("bond:Slippage")}</p>
+                        <p className="input-lable">
+                            <>{t('bond:Slippage')}</>
+                        </p>
                     </InputLabel>
                     <FormControl variant="outlined" color="primary" fullWidth>
                         <OutlinedInput
@@ -57,17 +65,23 @@ function AdvancedSettings({ open, handleClose, slippage, recipientAddress, onRec
                             }
                         />
                         <div className="help-text">
-                            <p className="text-bond-desc">{t("bond:SlippageHelpText")}</p>
+                            <p className="text-bond-desc">
+                                <>{t('bond:SlippageHelpText')}</>
+                            </p>
                         </div>
                     </FormControl>
 
                     <InputLabel htmlFor="recipient">
-                        <p className="input-lable">{t("bond:RecipientAddress")}</p>
+                        <p className="input-lable">
+                            <>{t('bond:RecipientAddress')}</>
+                        </p>
                     </InputLabel>
                     <FormControl variant="outlined" color="primary" fullWidth>
                         <OutlinedInput className="bond-input" id="recipient" value={recipientAddress} onChange={onRecipientAddressChange} type="text" />
                         <div className="help-text">
-                            <p className="text-bond-desc">{t("bond:RecipientAddressHelpText")}</p>
+                            <p className="text-bond-desc">
+                                <>{t('bond:RecipientAddressHelpText')}</>
+                            </p>
                         </div>
                     </FormControl>
                 </Box>
