@@ -1,209 +1,106 @@
-import { useSelector } from "react-redux";
-import { Grid, Zoom } from "@material-ui/core";
-import { trim } from "../../helpers";
-import "./dashboard.scss";
-import { Skeleton } from "@material-ui/lab";
-import { IReduxState } from "../../store/slices/state.interface";
-import { IAppSlice } from "../../store/slices/app-slice";
+import { lazy, Suspense, useEffect } from 'react';
 
-import { useTranslation } from "react-i18next";
+import { Box, Grow } from '@mui/material';
+import Grid from '@mui/material/Grid';
+import { useTranslation } from 'react-i18next';
+import { useDispatch, useSelector } from 'react-redux';
+
+import Loading from 'components/Loader';
+import { theme } from 'constants/theme';
+import { useWeb3Context } from 'contexts/web3/web3.context';
+import { formatUSD, formatAPY } from 'helpers/price-units';
+import { selectAppLoading, selectFormattedReservePrice } from 'store/modules/app/app.selectors';
+import { selectFormattedTreasuryBalance } from 'store/modules/bonds/bonds.selector';
+import { getTreasuryBalance } from 'store/modules/bonds/bonds.thunks';
+import { selectMarketsLoading } from 'store/modules/markets/markets.selectors';
+import { selectFormattedMarketCap, selectStakingRewards, selectTVL, selectWSBASHPrice } from 'store/modules/metrics/metrics.selectors';
+import { selectFormattedIndex } from 'store/modules/stake/stake.selectors';
+
+import './dashboard.scss';
+
+const MenuMetric = lazy(() => import('components/Metrics/MenuMetric'));
 
 function Dashboard() {
     const { t } = useTranslation();
+    const dispatch = useDispatch();
 
-    const isAppLoading = useSelector<IReduxState, boolean>(state => state.app.loading);
-    const app = useSelector<IReduxState, IAppSlice>(state => state.app);
+    const {
+        state: { networkID },
+    } = useWeb3Context();
 
-    const trimmedStakingAPY = trim(app.stakingAPY * 100, 1);
+    const marketsLoading = useSelector(selectMarketsLoading);
+    const bashPrice = useSelector(selectFormattedReservePrice);
+    const wsPrice = useSelector(selectWSBASHPrice);
+    const marketCap = useSelector(selectFormattedMarketCap);
+    const stakingRewards = useSelector(selectStakingRewards);
+    const TVL = useSelector(selectTVL);
+    const currentIndex = useSelector(selectFormattedIndex);
+
+    const treasuryBalance = useSelector(selectFormattedTreasuryBalance);
+    const appIsLoading = useSelector(selectAppLoading);
+
+    useEffect(() => {
+        if (!appIsLoading && networkID) {
+            dispatch(getTreasuryBalance({ networkID }));
+        }
+    }, [networkID, appIsLoading]);
+
+    if (appIsLoading) return <Loading />;
+
+    const APYMetrics = stakingRewards
+        ? [
+              { name: 'APY', value: stakingRewards ? formatAPY(stakingRewards.stakingAPY.toString()) : null },
+              { name: 'CurrentIndex', value: currentIndex },
+              { name: 'wsBASHPrice', value: wsPrice },
+          ]
+        : [];
+
+    const DashboardItems = [
+        { name: 'BashPrice', value: bashPrice },
+        { name: 'MarketCap', value: marketCap },
+        { name: 'TVL', value: formatUSD(TVL || 0, 2) },
+        { name: 'TreasuryBalance', value: treasuryBalance },
+
+        ...APYMetrics,
+        // { name: "RiskFreeValue", value: formatUSD(app.rfv) },
+        // { name: "RiskFreeValuewsBASH", value: formatUSD(app.rfv * Number(app.currentIndex)) },
+        // { name: "Runway", value: `${formatNumber(Number(app.runway), 1)} Days` },
+    ];
 
     return (
-        <div className="dashboard-view">
-            <div className="dashboard-infos-wrap">
-                <Zoom in={true}>
-                    <Grid container spacing={4}>
+        <Box>
+            <Grid container spacing={6} sx={{ p: 2 }} justifyContent="space-around">
+                {DashboardItems.map(metric => (
+                    <Grow in={!marketsLoading} {...(!marketsLoading ? { timeout: 1000 } : {})} key={`dashboard-item-${metric.name}`}>
                         <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("RiskFreeValue")}</p>
-                                <p className="card-value">
-                                    {isAppLoading ? (
-                                        <Skeleton width="250px" />
-                                    ) : (
-                                        new Intl.NumberFormat("en-US", {
-                                            style: "currency",
-                                            currency: "USD",
-                                            maximumFractionDigits: 2,
-                                            minimumFractionDigits: 2,
-                                        }).format(app.rfv)
-                                    )}
-                                </p>
-                                {/* <p className="card-value">
-                                    {isAppLoading ? (
-                                        <Skeleton width="250px" />
-                                    ) : (
-                                        new Intl.NumberFormat("en-US", {
-                                            style: "currency",
-                                            currency: "USD",
-                                            maximumFractionDigits: 2,
-                                            minimumFractionDigits: 2,
-                                        }).format(app.redeemRfv)
-                                    )}
-                                </p> */}
-                            </div>
+                            <Box
+                                className="Dashboard__box__item"
+                                sx={{
+                                    backgroundColor: theme.palette.cardBackground.main,
+                                    backdropFilter: 'blur(100px)',
+                                    borderRadius: '.5rem',
+                                    color: theme.palette.primary.main,
+                                    px: theme.spacing(2),
+                                    py: theme.spacing(4),
+                                    textAlign: 'center',
+                                    justifyContent: 'center',
+                                    alignItems: 'center',
+                                    display: 'flex',
+                                    flex: '1 1 auto',
+                                    overflow: 'auto',
+                                    flexDirection: 'column',
+                                    height: '100%',
+                                }}
+                            >
+                                <Suspense fallback={<Loading />}>
+                                    <MenuMetric metricKey={t(metric.name)} value={metric.value} />
+                                </Suspense>
+                            </Box>
                         </Grid>
-                        <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("RiskFreeValuewsBASH")}</p>
-                                <p className="card-value">
-                                    {isAppLoading ? (
-                                        <Skeleton width="250px" />
-                                    ) : (
-                                        new Intl.NumberFormat("en-US", {
-                                            style: "currency",
-                                            currency: "USD",
-                                            maximumFractionDigits: 2,
-                                            minimumFractionDigits: 2,
-                                        }).format(app.rfv * Number(app.currentIndex))
-                                    )}
-                                </p>
-                            </div>
-                        </Grid>
-                        {/* <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("RiskFreeValueDelta")}</p>
-                                <p className="card-value">{isAppLoading ? <Skeleton width="250px" /> : `${trim(Number(app.deltaMarketPriceRfv), 2)}%`}</p>
-                            </div>
-                        </Grid> */}
-                        <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("BASHPrice")}</p>
-                                <p className="card-value">{isAppLoading ? <Skeleton width="100px" /> : `$${trim(app.marketPrice, 2)}`}</p>
-                            </div>
-                        </Grid>
-
-                        <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("wsBASHPrice")}</p>
-                                <p className="card-value">{isAppLoading ? <Skeleton width="100px" /> : `$${trim(app.marketPrice * Number(app.currentIndex), 2)}`}</p>
-                            </div>
-                        </Grid>
-
-                        <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("MarketCap")}</p>
-                                <p className="card-value">
-                                    {isAppLoading ? (
-                                        <Skeleton width="160px" />
-                                    ) : (
-                                        new Intl.NumberFormat("en-US", {
-                                            style: "currency",
-                                            currency: "USD",
-                                            maximumFractionDigits: 0,
-                                            minimumFractionDigits: 0,
-                                        }).format(app.marketCap)
-                                    )}
-                                </p>
-                            </div>
-                        </Grid>
-
-                        {/* <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("SupplyStakedTotal")}</p>
-                                <p className="card-value">
-                                    {isAppLoading ? (
-                                        <Skeleton width="250px" />
-                                    ) : (
-                                        `${new Intl.NumberFormat("en-US", {
-                                            maximumFractionDigits: 0,
-                                            minimumFractionDigits: 0,
-                                        }).format(app.circSupply)}
-                                        /
-                                        ${new Intl.NumberFormat("en-US", {
-                                            maximumFractionDigits: 0,
-                                            minimumFractionDigits: 0,
-                                        }).format(app.totalSupply)}`
-                                    )}
-                                </p>
-                            </div>
-                        </Grid> */}
-
-                        <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("TVL")}</p>
-                                <p className="card-value">
-                                    {isAppLoading ? (
-                                        <Skeleton width="250px" />
-                                    ) : (
-                                        new Intl.NumberFormat("en-US", {
-                                            style: "currency",
-                                            currency: "USD",
-                                            maximumFractionDigits: 0,
-                                            minimumFractionDigits: 0,
-                                        }).format(app.stakingTVL)
-                                    )}
-                                </p>
-                            </div>
-                        </Grid>
-
-                        <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("APY")}</p>
-                                <p className="card-value">{isAppLoading ? <Skeleton width="250px" /> : `${new Intl.NumberFormat("en-US").format(Number(trimmedStakingAPY))}%`}</p>
-                            </div>
-                        </Grid>
-
-                        <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("CurrentIndex")}</p>
-                                <p className="card-value">{isAppLoading ? <Skeleton width="250px" /> : `${trim(Number(app.currentIndex), 2)} BASH`}</p>
-                            </div>
-                        </Grid>
-
-                        <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("TreasuryBalance")}</p>
-                                <p className="card-value">
-                                    {isAppLoading ? (
-                                        <Skeleton width="250px" />
-                                    ) : (
-                                        new Intl.NumberFormat("en-US", {
-                                            style: "currency",
-                                            currency: "USD",
-                                            maximumFractionDigits: 0,
-                                            minimumFractionDigits: 0,
-                                        }).format(app.treasuryBalance)
-                                    )}
-                                </p>
-                            </div>
-                        </Grid>
-
-                        {/* <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("BackinPerBASH")}</p>
-                                <p className="card-value">
-                                    {isAppLoading ? (
-                                        <Skeleton width="250px" />
-                                    ) : (
-                                        new Intl.NumberFormat("en-US", {
-                                            style: "currency",
-                                            currency: "USD",
-                                            maximumFractionDigits: 0,
-                                            minimumFractionDigits: 0,
-                                        }).format(app.rfv)
-                                    )}
-                                </p>
-                            </div>
-                        </Grid> */}
-
-                        <Grid item lg={6} md={6} sm={6} xs={12}>
-                            <div className="dashboard-card">
-                                <p className="card-title">{t("Runway")}</p>
-                                <p className="card-value">{isAppLoading ? <Skeleton width="250px" /> : `${trim(Number(app.runway), 1)} Days`}</p>
-                            </div>
-                        </Grid>
-                    </Grid>
-                </Zoom>
-            </div>
-        </div>
+                    </Grow>
+                ))}
+            </Grid>
+        </Box>
     );
 }
 
