@@ -1,8 +1,7 @@
-export { default as Bond } from "./Bond";
-export { default as ChooseBond } from "./ChooseBond";
-export { default as Stake } from "./Stake";
-export { default as Redeem } from "./Redeem";
-export { default as Forecast } from "./Forecast";
-export { default as Dashboard } from "./Dashboard";
-export { default as NotFound } from "./404";
-export { default as Wrap } from "./Wrap";
+export { default as Stake } from './Staking';
+export { default as Redeem } from './Redeem';
+export { default as Forecast } from './Forecast';
+export { default as Dashboard } from './Dashboard';
+export { default as NotFound } from './404';
+export { default as Wrap } from './Wrap';
+export { default as CritialError } from './CriticalError';
